// src/api/generated/models/loginResponse.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * BFF Service
 * Backend For Frontend service aggregating microservices
 * OpenAPI spec version: 1.0.0
 */

/**
 * Response model for successful login.
 */
export interface LoginResponse {
  access_token: string;
  id_token: string;
  refresh_token: string;
  token_type?: string;
  expires_in: number;
  user_groups?: string[];
}